requirements.txt
# requirements.txt
fredapi==2.1.0
pandas==2.1.0
python-dotenv==1.0.0
requests==2.31.0
jupyter==1.0.0
matplotlib==3.8.0
seaborn==0.12.0